+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0]
+
+### Fixed
+
+- some typos: also in settings
+
+## [0.11.0]
+
+### Added
+
+- custom capability: get default settings
+- custom capability: change settings
+
 ## [0.10.0]
 
 ### Added